SCI-9528 Simplify Syntax in Savu Plugins

+++ savu/plugins/filters/denoising/ccpi_denoising_cpu_3D_tools.py
@@ -45,7 +45,7 @@ efficient 2D/3D denoising.
 
         time_step:
              visibility: advanced
-             dtype: int
+             dtype: float
              description: 'Time marching step, relevant for ROF_TV, LLT_ROF,
                NDF, Diff4th methods.'
              default: 0.001
