+++ savu/plugins/filters/denoising/ccpi_denoising_gpu_tools.py
@@ -48,7 +48,7 @@ efficient 2D/3D denoising.
 
         time_step:
              visibility: advanced
-             dtype: int
+             dtype: float
              description: 'Time marching step, relevant for ROF_TV, LLT_ROF,
                NDF, Diff4th methods.'
              default: 0.001
